H8SCM: Add a type cast for printk to correct a warning message
Copy a type cast from the other cases of the same switch statement
to eliminate compiler warning messages.

+++ src/mainboard/supermicro/h8scm/agesawrapper.c
@@ -736,9 +736,9 @@ static void agesa_warning(EVENT_PARAMS *event)
 
 		case HT_EVENT_OPT_REQUIRED_CAP_RETRY:
 			printk(BIOS_DEBUG, "HT_EVENT_OPT_REQUIRED_CAP_RETRY, Socket %x Link %x Depth %x\n",
-				event->DataParam1,
-				event->DataParam2,
-				event->DataParam3);
+					(unsigned int)event->DataParam1,
+					(unsigned int)event->DataParam2,
+					(unsigned int)event->DataParam3);
 			break;
 
 		case HT_EVENT_OPT_REQUIRED_CAP_GEN3:
